==> 04 analisis secretoma.Rmd <==
---
title: "analisis secretoma"
output:
  html_document:
    df_print: paged
---

# librerias

```{r include=F}
library(tidyverse)
library(venneuler)
library(pdftools)
```

# Tablas

Se usan las tablas preparadas en el paso 02 combinando archivos.Rmd

```{r}
# Altamente transcritos Y diferencialmente expresados
HT_y_DGE <- read.csv("../00 resultados/seleccionados/HT y DGE/00 SOM con todo.txt", sep = "\t", header = T)

# solo diferencialmente expresados
DGE <- read.csv("../00 resultados/seleccionados/DGE/00 SOM con todo.txt", sep = "\t", header = T)

# solo altamente transcrito
HT <- read.csv("../00 resultados/seleccionados/HT/00 SOM con todo.txt", sep = "\t", header = T)

# proteínas HT y DGE de nodos no HT ni DGE

proteinas_sueltas <- read.csv("../00 resultados/seleccionados/proteinas restantes/00 SOM con todo.txt", sep = "\t", header = T)

# proteínas secretadas
secretadas <- read.csv("../00 resultados/seleccionados/proteinas secretadas/00 SOM con todo.txt", sep = "\t", header = T)
```

```{r}
preparando_analisis <- function(tabla) {
  # Se seleccionan las columnas de interés para el análisis secretómico: 
  tabla<- tabla%>% select(c("orden", "node", "proteinid", "secretoma_teorico", "CAZyme_function", "CAZymes", "C6_PSMs", "Wheat6_PSMs", "C14_PSMs", "Wheat14_PSMs", "familia_KOG", "KEGG_Definition", "funciones_GO", "funciones_IPR", "Pathway", "Pathway_class", "funciones_KOG"))
    
  # Se nombran con el nombre definitivo
  colnames(tabla) <- c("Order", "Node", "Protein-ID", "Secretome_by_Secretool", "Class", "Function", "Day_6_Control", "Day_6_Wheat", "Day_14_Control", "Day_14_Wheat",	"familia_KOG", "KEGG_Definition", "GO_function", "IPR_function",	"Pathway",	"Pathway_class",	"funciones_KOG")
  
  tabla$Secretome_by_Secretool <- gsub("YES", "+", tabla$Secretome_by_Secretool)
  tabla$Secretome_by_Secretool <- gsub("NO", "-", tabla$Secretome_by_Secretool)
  
 
tabla$Class <- ifelse(tabla$Class == "PCWDE", "CAZy (PCWDE)", 
                 ifelse(tabla$Class == "FCWDE", "CAZy (FCWDE)",
                    ifelse(tabla$Class == "PCWDE, FCWDE", "CAZy (PCWDE/FCWDE)",
                         ifelse(tabla$Class == "" & tabla$Function != "", "CAZy", ""))))
# Se quitan las medias por nodo, (se pone el if para evitar un problema con las proteínas que superan criterios dentro de los nodos meh, debido a que ahí no hay medias de nodos)
if (sum(grep("MEDIA", tabla$Node))!=0) {
  tabla <- tabla[-grep("MEDIA", tabla$Node),]  }
  tabla[,7:10] <- apply(tabla[,7:10], 2, function(x){as.numeric(as.character(x))})
# a partir de esta tabla se obtienen 2 tablas, una con las proteínas secretadas en el control y otra con las proteínas secretadas en el tratamiento.
  paja <- tabla[tabla$Day_6_Wheat>0 | tabla$Day_14_Wheat>0,] %>% select(-c("Day_6_Control","Day_14_Control"))
  
  control <- tabla[tabla$Day_6_Control>0 | tabla$Day_14_Control>0,] %>% select(-c("Day_6_Wheat","Day_14_Wheat"))
# las devuelve como una lista
return(list(paja, control))
}
```

```{r}
# Se aplica la función preparada previamente para cada tabla
HT_y_DGE <- preparando_analisis(HT_y_DGE)
HT <- preparando_analisis(HT)
DGE <- preparando_analisis(DGE)
proteinas_sueltas <- preparando_analisis(proteinas_sueltas)
secretadas <- preparando_analisis(secretadas)
```

Se guarda cada tabla

```{r}
write.table(HT_y_DGE[1], "../00 resultados/seleccionados/HT y DGE/01 secretados paja.txt", sep = "\t", quote = F, row.names = F)

write.table(HT_y_DGE[2], "../00 resultados/seleccionados/HT y DGE/01 secretados control.txt", sep = "\t", quote = F, row.names = F)

write.table(DGE[1], "../00 resultados/seleccionados/DGE/01 secretados paja.txt", sep = "\t", quote = F, row.names = F)
write.table(DGE[2], "../00 resultados/seleccionados/DGE/01 secretados control.txt", sep = "\t", quote = F, row.names = F)

write.table(HT[1], "../00 resultados/seleccionados/HT/01 secretados paja.txt", sep = "\t", quote = F, row.names = F)
write.table(HT[2], "../00 resultados/seleccionados/HT/01 secretados control.txt", sep = "\t", quote = F, row.names = F)
```

Y se combinan todas las tablas de interés en un único excel:

```{r}
library(openxlsx)

# el SOM entero
SOM <- read.csv("../00 resultados/00 SOM.txt", sep = "\t", header = T)

# Altamentes transcritos Y diferencialmente expresados
HT_y_DGE_nativo <- read.csv("../00 resultados/seleccionados/HT y DGE/00 SOM con todo.txt", sep = "\t", header = T)

# solo diferencialmente expresados
DGE_nativo <- read.csv("../00 resultados/seleccionados/DGE/00 SOM con todo.txt", sep = "\t", header = T)

# solo altamente transcrito
HT_nativo <- read.csv("../00 resultados/seleccionados/HT/00 SOM con todo.txt", sep = "\t", header = T)

# proteínas HT y DGE de nodos no HT ni DGE
proteinas_sueltas_nativas <- read.csv("../00 resultados/seleccionados/proteinas restantes/00 SOM con todo.txt", sep = "\t", header = T)

# proteínas secretadas
secretadas_nativas <- read.csv("../00 resultados/seleccionados/proteinas secretadas/00 SOM con todo.txt", sep = "\t", header = T)
```

Se combinan todas las tablas en una única lista

```{r}
tablas <- list("SOM" = SOM, "Highly Transcribed" = HT_nativo, "HT wheat" = HT[[1]], "HT Control" = HT[[2]], "DGE" =DGE_nativo, "DGE wheat" = DGE[[1]], "DGE Control" = DGE[[2]], "HT & DGE" = HT_y_DGE_nativo, "HT & DGE wheat" = HT_y_DGE[[1]], "HT & DGE Control"= HT_y_DGE[[2]], "other proteins"=proteinas_sueltas_nativas, "others wheat"= proteinas_sueltas[[1]], "others Control"= proteinas_sueltas[[2]], "Secreted"= secretadas_nativas, "Secreted wheat"= secretadas[[1]], "Secreted Control"= secretadas[[2]])
```

A partir de esa lista se construye el archivo de excel con todas las tablas separadas por pestañas.

```{r}
OUT <- createWorkbook(creator = "Rash")
# write.xlsx(tablas[1], file="../00 resultados/0000 Base de datos SOM.xlsx", sheetName=names(tablas)[1], row.names=FALSE, append = F)
for (i in 1:length(tablas)){
   
  addWorksheet(OUT, names(tablas)[i])
  writeData(OUT, sheet = names(tablas)[i], x = (tablas)[[i]], colNames = T, withFilter = F)


}
saveWorkbook(OUT,"../00 resultados/0000 Base de datos SOM.xlsx", overwrite = T)
```
